requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
google-generativeai==0.3.1
scikit-learn==0.24.2
joblib==1.0.1
numpy==1.20.3
pillow==9.0.0 